refactored to make easier to merge with refactor
